Add Language colors for the repositories.
Used colors.json from https://github.com/ozh/github-colors/blob/master/colors.json
to get the latest colors for each of the languages and then added to the repo-card
with the language and a circle for the color.

Add link around the Language to be the same as the Stars and forks.

Change the order of the language to match GitHub.